% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animove.R
\docType{data}
\name{buffalo_env}
\alias{buffalo_env}
\title{Environmental layers corresponding to the buffalo data set}
\usage{
data(buffalo_env)
}
\value{
RasterBrick with five layers:
\itemize{
\item elevation
\item slope
\item distance to water
\item mean NDVI
\item variance NDVI
}
}
\description{
Environmental layers corresponding to the buffalo data set
}
\examples{
data(buffalo_ll)
data(buffalo_utm)
data(buffalo_env)

plot(buffalo_env, 1)
lines(buffalo_utm)
}
\seealso{
buffalo_utm buffalo_ll
}
\keyword{datasets}
