% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animove.R
\docType{data}
\name{buffalo_env}
\alias{buffalo_env}
\title{Environmental layers coorresponding to the buffalo data set}
\usage{
data(buffalo_env)
}
\description{
Environmental layers coorresponding to the buffalo data set
}
\keyword{datasets}
